Added Database with Implementation, Profile Querying, and Increased Authentication (#33)
* Test

* Fixed Pass

* Test

* Fixed Pass

* Committing Styling for Landing and Home Pages

* Added styling to login and signup as well as fixed login bug

* Small button changes

* Started adding cards components

* Committing Cards and Deck with layout improvements

* Updated styling to be MAC compatible

* Refactored Board and Front End Responsibilities

Too much of the game page front end relied on the board component. I moved this over so that the board is actually just a component of the greater game handler.

* Resized the card images

* Committing Database with increased site authentication

Database is up and running on firebase and Profile now displays some basic information from the database. More to come.

---------

Co-authored-by: TyKerkman <tkermkan3@huskers.unl.edu>